Move TaskList into tasks package
