+++ src/main/java/sunderray/commands/AddCommand.java
@@ -1,7 +1,7 @@
 package sunderray.commands;
 
 import sunderray.data.messages.InfoMsg;
-import sunderray.tasklist.TaskList;
+import sunderray.tasks.TaskList;
 import sunderray.tasks.Task;
 
 /**
