+++ src/main/java/sunderray/commands/FindCommand.java
@@ -1,7 +1,7 @@
 package sunderray.commands;
 
 import sunderray.data.messages.InfoMsg;
-import sunderray.tasklist.TaskList;
+import sunderray.tasks.TaskList;
 
 /**
  * Finds tasks whose description matches the keyword.
